Migrate from CircleCI to GitHub Actions (#17)
The old obsolete CircleCI config has been replaced by the standard
GitHub Actions workflow we use for testing libraries.

Now Node.js versions v14 and v16 are also tested.

The test environment was switched to Headless Chrome and Headless
Firefox, from the non-headless versions. This makes testing on CI much
simpler.

The test configuration was also simplified to remove the global install
commands and the instruction to use Node.js v6. Now the local version
of Node.js is used (v12), and the normal development dependencies are
installed before the test is run.