Add ESLint using standard config (#16)
The standard ESLint config has been applied, with all lint errors
fixed. Most of the errors could be fixed automatically.

TypeScript was updated to the latest v4.0 patch because it was a
requirement of our ESLint TypeScript config. Also the README example
was improved slightly (it now uses `const` instead of `var`, and
includes the `assert` import that was missing).