Add comprehensive unit tests (#21)
Unit tests have been added to test decrypting data that was encrypted
using an older version of this library. Previously encryption and
decryption were only tested together, so if a breaking change was made
to both encryption and decryption, it may have been missed. Now we
should have some guarantee of stability.

Tests have been added for the other untested exported functions as
well, which are 'keyFromPassword', 'encryptWithKey', 'decryptWithKey',
and 'generateSalt'.

The ESLint configuration was updated to state that these tests are
expected to run in a browser, because the `atob` API was required for
testing `generateSalt`.