Remove `browserify-unibabel` (#13)
This project was using `browserify-unibabel` rather than `Buffer` to
perform certain low-level operations. Presumably this was done to
ensure this package works out of the box in the browser, as `Buffer` is
Node.js- specific.

However, the npm package `buffer` already accomplishes this, and it's
used by default by `browserify` and most other bundlers. So in effect
this package is already web-friendly even when using `Buffer` directly,
as long as the consumer uses a bundler. Moreover, this package already
relied upon the use of a bundler because `browserify-unibabel` uses
CommonJS. It was created for use with `browserify` (hence the name).

All uses of `unibabel` have been replaced by the equivalent `Buffer`
operation. This should work exactly the same except with one less
dependency.

The `types` directory was removed from the list of ESLint ignore
patterns, as that directory no longer exists.