Add `@metamask/auto-changelog` (#28)
`@metamask/auto-changelog` has been added to help with updating and
validating the changelog. A basic empty changelog has been added as
well, along with changelog validation in CI.

The `repository` field of `package.json` was updated to use an HTTPS
URL, as that's the only URL format understood by
`@metamask/auto-changelog`.