Switch from `testem`/`QUnit` to Playwright
Playwright is now used to run unit tests in Chromium and Firefox, which
are our two supported browsers. This resolves issues we've had with
using `testem` (it wasn't finding tests in CI, it wasn't running
Chromium tests locally, it didn't seem to be reading the configuration
file or running the pre-test command, etc).

Playwright on the other hand is well maintained and easier to debug.