Use standard exports (#25)
The syntax previously used to export functions was preventing this
module from being imported from another TypeScript file. The idiomatic
way of exporting named functions is to prefix the function declaration
with `export`. This seems to result in functionally equivalent output,
and it is easier to import from TypeScript now.